edit-feat: CartController.php , CartItemController.php , Cart.php ,CartItemService.php , create_carts_table , routes/api.php <Some updates>
