+++ app/Http/Controllers/CartItem/CartItemController.php
@@ -6,8 +6,6 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\CartItem\StoreCartItemRequest;
 use App\Http\Requests\CartItem\UpdateCartItemRequest;
 use App\Http\Resources\CartItemResource;
-use App\Http\Resources\CartResource;
-use App\Models\Cart\Cart;
 use App\Models\CartItem\CartItem;
 use App\Services\CartItem\CartItemService;
 
@@ -21,18 +19,6 @@ class CartItemController extends Controller
         $this->cartItemService = $cartItemService;
     }
 
-    /**
-     * View all user carts for ((Admin))
-     *
-     * @return \Illuminate\Http\JsonResponse
-     */
-
-    public function index()
-    {
-        $data = Cart::with(['cartItems.product'])->get();
-        return self::success(CartResource::collection($data));
-
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -49,16 +35,6 @@ class CartItemController extends Controller
         return self::success(null, 'Added successfully!', 201);
     }
 
-    /**
-     *   Display the specified cart for ((Admin))
-     *
-     * @param Cart $cartItem
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function show(Cart $cartItem)
-    {
-        return self::success(new CartResource($cartItem->load(['user', 'cartItems.product'])));
-    }
 
     /**
      *  Update the specified resource in storage.
@@ -81,18 +57,9 @@ class CartItemController extends Controller
      */
     public function destroy(CartItem $cartItem)
     {
-        $cartItem->forceDelete();
+        $this->cartItemService->deleteItem($cartItem);
         return self::success(null, 'deleted successfully!');
     }
 
-    /**
-     *  View the cart of the auth user.
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function userCart()
-    {
-        $cart = Cart::where('user_id', auth()->user()->id)->with('cartItems.product')->first();
-        return self::success(new CartResource($cart));
-    }
 
 }
