+++ app/Models/Cart/Cart.php
@@ -10,7 +10,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Cart extends Model
 {
-    use HasFactory, SoftDeletes;
+    use HasFactory;
 
     /**
      * The attributes that are mass assignable.
@@ -30,7 +30,7 @@ class Cart extends Model
 
     /**
      * Get the user associated with the cart.
-     * 
+     *
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
     public function user()
@@ -40,7 +40,7 @@ class Cart extends Model
 
     /**
      * Get the cart items associated with the cart.
-     * 
+     *
      * @return \Illuminate\Database\Eloquent\Relations\HasMany
      */
     public function cartItems()
