+++ routes/api.php
@@ -1,21 +1,20 @@
 <?php
 
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Cart\CartController;
+use App\Http\Controllers\CartItem\CartItemController;
+use App\Http\Controllers\Category\MainCategoryController;
+use App\Http\Controllers\Category\SubCategoryController;
+use App\Http\Controllers\Favorite\FavoriteController;
+use App\Http\Controllers\Order\OrderController;
+use App\Http\Controllers\Permission\PermissionController;
+use App\Http\Controllers\Photo\PhotoController;
+use App\Http\Controllers\Product\ProductController;
 use App\Http\Controllers\Rate\RateController;
 use App\Http\Controllers\Role\RoleController;
 use App\Http\Controllers\User\AuthController;
-use App\Http\Controllers\User\UserController;
-use App\Http\Controllers\Order\OrderController;
-use App\Http\Controllers\Photo\PhotoController;
-use App\Http\Controllers\Product\ProductController;
-use App\Http\Controllers\CartItem\CartItemController;
-use App\Http\Controllers\Favorite\FavoriteController;
 use App\Http\Controllers\User\PasswordResetController;
-use App\Http\Controllers\Category\SubCategoryController;
-use App\Http\Controllers\Category\MainCategoryController;
-use App\Http\Controllers\Permission\PermissionController;
+use App\Http\Controllers\User\UserController;
+use Illuminate\Support\Facades\Route;
 
 /*
 |--------------------------------------------------------------------------
@@ -34,13 +33,15 @@ Route::post('/login', [AuthController::class, 'login']);
 //Oauth
 Route::get('/auth/{provider}', [AuthController::class, 'redirectToProvider']);
 Route::get('/auth/{provider}/callback', [AuthController::class, 'handleProviderCallback']);
+
 Route::middleware('auth:api')->group(function () {
     Route::post('/logout', [AuthController::class, 'logout']);
     Route::post('/refresh-token', [AuthController::class, 'refresh']);
 
     //Cart & Cart item-----------------------------------------------------------
-    Route::apiResource('/cart-items', CartItemController::class);
-    Route::get('/user-cart', [CartItemController::class, 'userCart']);
+    Route::apiResource('/cart-items', CartItemController::class)->except(['index', 'show']);
+    Route::apiResource('/carts', CartController::class)->only(['index', 'show']);
+    Route::get('/user-cart', [CartController::class, 'userCart']);
 
 });
